Extract `prepass_bindings.rs`
